+++ app.js
@@ -4,22 +4,16 @@ const mongoose = require("mongoose");
 
 const cookieParser = require("cookie-parser");
 
-const { celebrate, Joi, errors } = require("celebrate");
+const { errors } = require("celebrate");
 
 const cors = require("cors");
 
-const { auth } = require("./middlewares/auth");
-
-const { login, logout, signup } = require("./controllers/userControllers");
 const { errorHandler } = require("./middlewares/errorHandler");
 const { requestLogger, errorLogger } = require("./middlewares/logger");
-const { NotFoundError } = require("./errors/NotFoundError");
-
-const userRoutes = require("./routes/userRouter");
-const movieRoutes = require("./routes/movieRouter");
+const routes = require("./routes/index");
 
 const { PORT = 3000 } = process.env;
-mongoose.connect("mongodb://127.0.0.1:27017/filmsdb");
+mongoose.connect(process.env.NODE_ENV === "production" ? process.env.DATA_BASE : "mongodb://127.0.0.1:27017/filmsdb");
 
 const app = express();
 
@@ -35,30 +29,8 @@ app.use(cors(corsOptions));
 app.use(cookieParser());
 app.use(requestLogger);
 
-app.post("/signin", celebrate({
-  body: Joi.object().keys({
-    email: Joi.string().required().email(),
-    password: Joi.string().required(),
-  }),
-}), login);
+app.use(routes);
 
-app.post("/signup", celebrate({
-  body: Joi.object().keys({
-    email: Joi.string().required().email(),
-    password: Joi.string().required(),
-    name: Joi.string().min(2).max(30),
-  }),
-}), signup);
-
-app.post("/signout", logout);
-
-app.use(auth);
-app.use("/users", userRoutes);
-app.use("/movies", movieRoutes);
-
-app.use((req, res, next) => {
-  next(new NotFoundError("Страница не найдена"));
-});
 app.use(errorLogger);
 app.use(errors());
 app.use(errorHandler);
